update: panel config save/load
