+++ console/packages/starwhale-core/src/widget/index.tsx
@@ -1,6 +1,4 @@
 export { default as BaseWidget } from './BaseWidget'
-export * from './hooks/useWidget'
-
 export { default as WidgetFactory } from './WidgetFactory'
 export * from './WidgetFactory'
 export { default as WidgetPlugin } from './WidgetPlugin'
